API PYTHON
In this study I seek to understand in a simplified way the development of an API.

This program was developed in Python language. The development was through the Replit platform (responsible for creating the API host). In the api.py document, we observe the construction of the API functionality, being used again to present a sales table to the user - as shown in the import_api.py file